# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/repowiki/by-project/{projectId}": {
            "get": {
                "description": "根据关联项目 ID 查询对应的 RepoWiki 配置，未创建配置时返回空数据（前端据此判断\"不存在→创建\"）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RepoWiki接口"
                ],
                "summary": "[管理] 按项目ID查询 RepoWiki 配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功（data 为 null 表示未创建配置）",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/repowiki.ConfigResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/repowiki/configs": {
            "get": {
                "description": "按 page/size 分页查询 RepoWiki 配置列表，返回配置信息与总数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RepoWiki接口"
                ],
                "summary": "[管理] 获取 RepoWiki 配置列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/repowiki.ConfigListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "提交仓库地址、分支、语言等参数创建 RepoWiki 分析配置，可选 SSH Key 与 Wiki 访问密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RepoWiki接口"
                ],
                "summary": "[管理] 创建 RepoWiki 配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建配置请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repowiki.CreateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/repowiki.ConfigResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/repowiki/configs/{id}/analyze": {
            "post": {
                "description": "根据配置 ID 触发 RepoWiki 仓库分析（异步执行），返回版本 ID 用于轮询状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RepoWiki接口"
                ],
                "summary": "[管理] 触发仓库分析",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "分析请求（语言/分支可选）",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/repowiki.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分析任务已触发",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/repowiki.AnalyzeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "配置不存在",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "分析任务已达并发上限",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "LLM Provider 未就绪",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        },
...
        "repowiki.ConfigListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "配置列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repowiki.ConfigResponse"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "repowiki.ConfigResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "default_branch": {
                    "description": "默认分支",
                    "type": "string"
                },
                "default_language": {
                    "description": "默认Wiki语言",
                    "type": "string"
                },
                "has_password": {
                    "description": "是否设置Wiki密码",
                    "type": "boolean"
                },
                "id": {
                    "description": "配置ID",
                    "type": "integer"
                },
                "last_accessed_at": {
                    "description": "最后访问时间",
                    "type": "string"
                },
                "latest_version": {
                    "description": "最近版本状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/repowiki.VersionStatusResponse"
                        }
                    ]
                },
                "name": {
                    "description": "配置名称",
                    "type": "string"
                },
                "project_id": {
                    "description": "关联项目ID",
                    "type": "integer"
                },
                "repo_url": {
                    "description": "Git仓库地址",
                    "type": "string"
                },
                "ssh_key_id": {
                    "description": "关联SSH密钥ID（字符串形式的雪花ID，nil 表示未关联）",
                    "type": "string"
                },
                "status": {
                    "description": "当前分析状态",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
    }
}