# internal/handler/repowiki.go
package handler

import (
	"strconv"

	xSnowflake "github.com/bamboo-services/bamboo-base-go/common/snowflake"
	xResult "github.com/bamboo-services/bamboo-base-go/major/result"
	"github.com/gin-gonic/gin"
	apiCommon "github.com/xiaolfeng/Lumina/api/common"
	apiRepowiki "github.com/xiaolfeng/Lumina/api/repowiki"
	"github.com/xiaolfeng/Lumina/internal/entity"
)

// 确保 apiCommon 包被编译器识别（swag 注释依赖此导入）
var _ = apiCommon.BaseResponse{}

// ──────────────────────────────────────────────────────────────────────
// RepoWikiHandler
// ──────────────────────────────────────────────────────────────────────

// CreateConfig 创建 RepoWiki 配置
//
// @Summary     [管理] 创建 RepoWiki 配置
// @Description 提交仓库地址、分支、语言等参数创建 RepoWiki 分析配置，可选 SSH Key 与 Wiki 访问密码
// @Tags        RepoWiki接口
// @Accept      json
// @Produce     json
// @Param       Authorization  header    string                              true  "Bearer Access Token"
// @Param       request        body      apiRepowiki.CreateConfigRequest     true  "创建配置请求"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.ConfigResponse}  "创建成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Router      /api/v1/repowiki/configs [POST]
func (h *RepoWikiHandler) CreateConfig(ctx *gin.Context) {
	h.log.Info(ctx, "CreateConfig - 创建 RepoWiki 配置")

	var req apiRepowiki.CreateConfigRequest
	if !BindJSON(ctx, &req) {
		return
	}

	config, xErr := h.service.repoWikiLogic.CreateConfig(ctx.Request.Context(), &req)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "创建成功", configToResponse(config, nil))
}

// ListConfigs 获取 RepoWiki 配置列表
//
// @Summary     [管理] 获取 RepoWiki 配置列表
// @Description 按 page/size 分页查询 RepoWiki 配置列表，返回配置信息与总数
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       page           query     int      false  "页码"
// @Param       size           query     int      false  "每页数量"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.ConfigListResponse}  "获取成功"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Router      /api/v1/repowiki/configs [GET]
func (h *RepoWikiHandler) ListConfigs(ctx *gin.Context) {
	h.log.Info(ctx, "ListConfigs - 获取 RepoWiki 配置列表")

	pageStr := ctx.DefaultQuery("page", "1")
	sizeStr := ctx.DefaultQuery("size", "20")
	page, _ := strconv.Atoi(pageStr)
	size, _ := strconv.Atoi(sizeStr)

	configs, total, xErr := h.service.repoWikiLogic.ListConfigs(ctx.Request.Context(), page, size)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	items := make([]apiRepowiki.ConfigResponse, 0, len(configs))
	for _, c := range configs {
		items = append(items, *configToResponse(c, nil))
	}

	xResult.SuccessHasData(ctx, "获取成功", apiRepowiki.ConfigListResponse{
		Total: total,
		Items: items,
	})
}

// GetConfig 获取 RepoWiki 配置详情
//
// @Summary     [管理] 获取 RepoWiki 配置详情
// @Description 根据配置 ID 查询单个 RepoWiki 配置详情
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "配置ID"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.ConfigResponse}  "获取成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Router      /api/v1/repowiki/configs/{id} [GET]
func (h *RepoWikiHandler) GetConfig(ctx *gin.Context) {
	h.log.Info(ctx, "GetConfig - 获取 RepoWiki 配置详情")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	config, xErr := h.service.repoWikiLogic.GetConfig(ctx.Request.Context(), id)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "获取成功", configToResponse(config, nil))
}

// GetConfigByProjectID 按项目 ID 查询 RepoWiki 配置
//
// @Summary     [管理] 按项目ID查询 RepoWiki 配置
// @Description 根据关联项目 ID 查询对应的 RepoWiki 配置，未创建配置时返回空数据（前端据此判断"不存在→创建"）
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       projectId      path      string   true  "项目ID"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.ConfigResponse}  "获取成功（data 为 null 表示未创建配置）"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Router      /api/v1/repowiki/by-project/{projectId} [GET]
func (h *RepoWikiHandler) GetConfigByProjectID(ctx *gin.Context) {
	h.log.Info(ctx, "GetConfigByProjectID - 按项目 ID 查询 RepoWiki 配置")

	projectID, xErr := ParseSnowflakeID(ctx, ctx.Param("projectId"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	config, found, xErr := h.service.repoWikiLogic.GetConfigByProjectID(ctx.Request.Context(), projectID)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}
	if !found {
		xResult.SuccessHasData(ctx, "未创建配置", nil)
		return
	}

	xResult.SuccessHasData(ctx, "获取成功", configToResponse(config, nil))
}

// UpdateConfig 更新 RepoWiki 配置
//
// @Summary     [管理] 更新 RepoWiki 配置
// @Description 更新指定 ID 的 RepoWiki 配置信息（支持部分更新，指针 nil = 不更新）
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "配置ID"
// @Param       request        body      apiRepowiki.UpdateConfigRequest     true  "更新配置请求"
// @Success     200  {object}  apiCommon.BaseResponse  "更新成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Router      /api/v1/repowiki/configs/{id} [PUT]
func (h *RepoWikiHandler) UpdateConfig(ctx *gin.Context) {
	h.log.Info(ctx, "UpdateConfig - 更新 RepoWiki 配置")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	var req apiRepowiki.UpdateConfigRequest
	if !BindJSON(ctx, &req) {
		return
	}

	xErr = h.service.repoWikiLogic.UpdateConfig(ctx.Request.Context(), id, &req)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.Success(ctx, "更新成功")
}

// DeleteConfig 删除 RepoWiki 配置
//
// @Summary     [管理] 删除 RepoWiki 配置
// @Description 根据配置 ID 删除指定 RepoWiki 配置，删除后不可恢复
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "配置ID"
// @Success     200  {object}  apiCommon.BaseResponse  "删除成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Router      /api/v1/repowiki/configs/{id} [DELETE]
func (h *RepoWikiHandler) DeleteConfig(ctx *gin.Context) {
	h.log.Info(ctx, "DeleteConfig - 删除 RepoWiki 配置")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xErr = h.service.repoWikiLogic.DeleteConfig(ctx.Request.Context(), id)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.Success(ctx, "删除成功")
}

// Analyze 触发仓库分析
//
// @Summary     [管理] 触发仓库分析
// @Description 根据配置 ID 触发 RepoWiki 仓库分析（异步执行），返回版本 ID 用于轮询状态
// @Tags        RepoWiki接口
// @Accept      json
// @Produce     json
// @Param       Authorization  header    string                        true  "Bearer Access Token"
// @Param       id             path      string                        true  "配置ID"
// @Param       request        body      apiRepowiki.AnalyzeRequest    false "分析请求（语言/分支可选）"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.AnalyzeResponse}  "分析任务已触发"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Failure     409  {object}  apiCommon.BaseResponse  "分析任务已达并发上限"
// @Failure     500  {object}  apiCommon.BaseResponse  "LLM Provider 未就绪"
// @Router      /api/v1/repowiki/configs/{id}/analyze [POST]
func (h *RepoWikiHandler) Analyze(ctx *gin.Context) {
	h.log.Info(ctx, "Analyze - 触发仓库分析")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	var req apiRepowiki.AnalyzeRequest
	// body 可选，绑定失败时使用零值（默认分支和语言由 Logic 层填充）
	_ = ctx.ShouldBindJSON(&req)

	version, xErr := h.service.repoWikiLogic.AnalyzeRepo(ctx.Request.Context(), id, &req)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "分析任务已触发", apiRepowiki.AnalyzeResponse{
		VersionID: version.ID.Int64(),
		Status:    version.Status,
	})
}

// Update 触发增量更新
//
// @Summary     [管理] 触发增量更新
// @Description 根据配置 ID 触发 RepoWiki 增量更新（异步执行），自动对比 commit hash 决定全量或增量分析
// @Tags        RepoWiki接口
// @Accept      json
// @Produce     json
// @Param       Authorization  header    string                        true  "Bearer Access Token"
// @Param       id             path      string                        true  "配置ID"
// @Param       request        body      apiRepowiki.AnalyzeRequest    false "更新请求（语言/分支可选）"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.AnalyzeResponse}  "更新任务已触发"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Failure     409  {object}  apiCommon.BaseResponse  "分析任务已达并发上限"
// @Failure     500  {object}  apiCommon.BaseResponse  "LLM Provider 未就绪"
// @Router      /api/v1/repowiki/configs/{id}/update [PUT]
func (h *RepoWikiHandler) Update(ctx *gin.Context) {
	h.log.Info(ctx, "Update - 触发增量更新")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	var req apiRepowiki.AnalyzeRequest
	_ = ctx.ShouldBindJSON(&req)

	version, xErr := h.service.repoWikiLogic.AnalyzeRepo(ctx.Request.Context(), id, &req)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "更新任务已触发", apiRepowiki.AnalyzeResponse{
		VersionID: version.ID.Int64(),
		Status:    version.Status,
	})
}

// GetVersionStatus 获取版本分析状态
//
// @Summary     [管理] 获取版本分析状态
// @Description 根据版本 ID 查询分析状态（pending/analyzing/completed/failed），用于轮询分析进度
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "版本ID"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.VersionStatusResponse}  "获取成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "版本不存在"
// @Router      /api/v1/repowiki/versions/{id}/status [GET]
func (h *RepoWikiHandler) GetVersionStatus(ctx *gin.Context) {
	h.log.Info(ctx, "GetVersionStatus - 获取版本分析状态")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	version, xErr := h.service.repoWikiLogic.GetVersionStatus(ctx.Request.Context(), id)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "获取成功", versionToStatusResponse(version))
}

// GetVersionDetail 获取版本详情
//
// @Summary     [管理] 获取版本详情
// @Description 根据版本 ID 查询完整的版本信息（含 commit hash、分析耗时、token 消耗等）
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "版本ID"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.VersionStatusResponse}  "获取成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "版本不存在"
// @Router      /api/v1/repowiki/versions/{id} [GET]
func (h *RepoWikiHandler) GetVersionDetail(ctx *gin.Context) {
	h.log.Info(ctx, "GetVersionDetail - 获取版本详情")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	version, xErr := h.service.repoWikiLogic.GetVersionStatus(ctx.Request.Context(), id)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "获取成功", versionToStatusResponse(version))
}

// ListVersions 获取版本列表
//
// @Summary     [管理] 获取版本列表
// @Description 按配置 ID 分页查询 Wiki 版本列表，返回版本状态与总数
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "配置ID"
// @Param       page           query     int      false  "页码"
// @Param       size           query     int      false  "每页数量"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.VersionListResponse}  "获取成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Router      /api/v1/repowiki/configs/{id}/versions [GET]
func (h *RepoWikiHandler) ListVersions(ctx *gin.Context) {
	h.log.Info(ctx, "ListVersions - 获取版本列表")

	configID, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	pageStr := ctx.DefaultQuery("page", "1")
	sizeStr := ctx.DefaultQuery("size", "20")
	page, _ := strconv.Atoi(pageStr)
	size, _ := strconv.Atoi(sizeStr)

	versions, total, xErr := h.service.repoWikiLogic.ListVersions(ctx.Request.Context(), configID, page, size)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	items := make([]apiRepowiki.VersionStatusResponse, 0, len(versions))
	for _, v := range versions {
		items = append(items, *versionToStatusResponse(v))
	}

	xResult.SuccessHasData(ctx, "获取成功", apiRepowiki.VersionListResponse{
		Total: total,
		Items: items,
	})
}

// ──────────────────────────────────────────────────────────────────────
// Entity → DTO 转换辅助函数
// ──────────────────────────────────────────────────────────────────────

// configToResponse 将 RepoWikiConfig 实体转为响应 DTO
//
// latestVersion 为 nil 时 ConfigResponse.LatestVersion 留空（omitempty）。
func configToResponse(config *entity.RepoWikiConfig, latestVersion *entity.WikiVersion) *apiRepowiki.ConfigResponse {
	resp := &apiRepowiki.ConfigResponse{
		ID:              config.ID.Int64(),
		ProjectID:       config.ProjectID.Int64(),
		Name:            config.Name,
		RepoURL:         config.GitURL,
		DefaultBranch:   config.DefaultBranch,
		DefaultLanguage: config.DefaultLanguage,
		Status:          config.Status,
		SSHKeyID:        sshKeyIDToString(config.SSHKeyID),
		HasPassword:     config.WikiPasswordHash != "",
		LastAccessedAt:  config.LastAccessedAt,
		CreatedAt:       config.CreatedAt,
		UpdatedAt:       config.UpdatedAt,
	}

	if latestVersion != nil {
		resp.LatestVersion = versionToStatusResponse(latestVersion)
	}

	return resp
}

// sshKeyIDToString 将 *xSnowflake.SnowflakeID 转为 *string（API 返回字符串形式的雪花 ID）
func sshKeyIDToString(id *xSnowflake.SnowflakeID) *string {
	if id == nil {
		return nil
	}
	s := strconv.FormatInt(id.Int64(), 10)
	return &s
}

// versionToStatusResponse 将 WikiVersion 实体转为版本状态响应 DTO
func versionToStatusResponse(version *entity.WikiVersion) *apiRepowiki.VersionStatusResponse {
	return &apiRepowiki.VersionStatusResponse{
		ID:              version.ID.Int64(),
		ConfigID:        version.ConfigID.Int64(),
		CommitHash:      version.CommitHash,
		Branch:          version.Branch,
		Language:        version.Language,
		Status:          version.Status,
		CurrentStage:    version.CurrentStage,
		ProgressPercent: 0, // 实体未存储进度百分比，Pipeline 阶段性更新通过 CurrentStage 体现
		ErrorMsg:        version.ErrorMsg,
		FileCount:       version.FileCount,
		TokenCount:      version.TokenCount,
		DurationMs:      version.DurationMs,
		StartedAt:       version.StartedAt,
		CompletedAt:     version.CompletedAt,
		CreatedAt:       version.CreatedAt,
	}
}

// GetWebhookConfig 获取 Webhook 配置
//
// @Summary     [管理] 获取 Webhook 配置
// @Description 根据配置 ID 查询 Webhook 地址、Token、Secret 状态和监控分支
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "配置ID"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.WebhookConfigResponse}  "获取成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Router      /api/v1/repowiki/configs/{id}/webhook [GET]
func (h *RepoWikiHandler) GetWebhookConfig(ctx *gin.Context) {
	h.log.Info(ctx, "GetWebhookConfig - 获取 Webhook 配置")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	info, xErr := h.service.repoWikiLogic.GetWebhookConfig(ctx.Request.Context(), id, ctx.Request.Host)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "获取成功", apiRepowiki.WebhookConfigResponse{
		URL:       info.URL,
		Token:     info.Token,
		HasSecret: info.HasSecret,
		Branches:  info.Branches,
	})
}

// UpdateWebhookBranches 更新 Webhook 监控分支
//
// @Summary     [管理] 更新 Webhook 监控分支
// @Description 更新指定配置监控的 Git 分支列表，仅列出的分支推送会触发 Wiki 分析
// @Tags        RepoWiki接口
// @Accept      json
// @Produce     json
// @Param       Authorization  header    string                                     true  "Bearer Access Token"
// @Param       id             path      string                                     true  "配置ID"
// @Param       request        body      apiRepowiki.UpdateWebhookBranchesRequest  true  "更新分支请求"
// @Success     200  {object}  apiCommon.BaseResponse  "更新成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Router      /api/v1/repowiki/configs/{id}/webhook/branches [PUT]
func (h *RepoWikiHandler) UpdateWebhookBranches(ctx *gin.Context) {
	h.log.Info(ctx, "UpdateWebhookBranches - 更新 Webhook 监控分支")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	var req apiRepowiki.UpdateWebhookBranchesRequest
	if !BindJSON(ctx, &req) {
		return
	}

	xErr = h.service.repoWikiLogic.UpdateWebhookBranches(ctx.Request.Context(), id, req.Branches)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.Success(ctx, "更新成功")
}

// RegenerateWebhook 重新生成 Webhook 凭据
//
// @Summary     [管理] 重新生成 Webhook 凭据
// @Description 为指定配置重新生成 Webhook Token 和 Secret，旧 Token 立即失效，仅本次返回完整凭据
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "配置ID"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.RegenerateWebhookResponse}  "重新生成成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Failure     404  {object}  apiCommon.BaseResponse  "配置不存在"
// @Router      /api/v1/repowiki/configs/{id}/webhook/regenerate [POST]
func (h *RepoWikiHandler) RegenerateWebhook(ctx *gin.Context) {
	h.log.Info(ctx, "RegenerateWebhook - 重新生成 Webhook 凭据")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	token, secret, xErr := h.service.repoWikiLogic.RegenerateWebhook(ctx.Request.Context(), id)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	// URL 由 GetWebhookConfig 提供，重新生成后需前端再次获取完整配置
	xResult.SuccessHasData(ctx, "重新生成成功", apiRepowiki.RegenerateWebhookResponse{
		Token:  token,
		Secret: secret,
	})
}

// ListWebhookEvents 获取 Webhook 事件列表
//
// @Summary     [管理] 获取 Webhook 事件列表
// @Description 根据配置 ID 分页查询 Webhook 接收事件及其处理结果
// @Tags        RepoWiki接口
// @Produce     json
// @Param       Authorization  header    string   true  "Bearer Access Token"
// @Param       id             path      string   true  "配置ID"
// @Param       page           query     int      false  "页码"
// @Param       size           query     int      false  "每页数量"
// @Success     200  {object}  apiCommon.BaseResponse{data=apiRepowiki.WebhookEventListResponse}  "获取成功"
// @Failure     400  {object}  apiCommon.BaseResponse  "请求参数错误"
// @Failure     401  {object}  apiCommon.BaseResponse  "未授权"
// @Router      /api/v1/repowiki/configs/{id}/webhook/events [GET]
func (h *RepoWikiHandler) ListWebhookEvents(ctx *gin.Context) {
	h.log.Info(ctx, "ListWebhookEvents - 获取 Webhook 事件列表")

	id, xErr := ParseSnowflakeID(ctx, ctx.Param("id"))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	pageStr := ctx.DefaultQuery("page", "1")
	sizeStr := ctx.DefaultQuery("size", "20")
	page, _ := strconv.Atoi(pageStr)
	size, _ := strconv.Atoi(sizeStr)

	events, total, xErr := h.service.repoWikiLogic.ListWebhookEvents(ctx.Request.Context(), id, page, size)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	items := make([]apiRepowiki.WebhookEventResponse, 0, len(events))
	for _, e := range events {
		items = append(items, webhookEventToResponse(e))
	}

	xResult.SuccessHasData(ctx, "获取成功", apiRepowiki.WebhookEventListResponse{
		Total: total,
		Items: items,
	})
}

// webhookEventToResponse 将 WebhookEvent 实体转为响应 DTO
func webhookEventToResponse(event *entity.WebhookEvent) apiRepowiki.WebhookEventResponse {
	resp := apiRepowiki.WebhookEventResponse{
		ID:           event.ID.Int64(),
		Provider:     event.Provider,
		EventType:    event.EventType,
		Branch:       event.Branch,
		CommitBefore: event.CommitBefore,
		CommitAfter:  event.CommitAfter,
		ChangedCount: event.ChangedCount,
		Status:       event.Status,
		Reason:       event.Reason,
		ResponseCode: event.ResponseCode,
		ReceivedAt:   event.ReceivedAt,
		ProcessedAt:  event.ProcessedAt,
	}

	if event.ConfigID != nil {
		resp.ConfigID = event.ConfigID.Int64()
	}
	if !event.VersionID.IsZero() {
		resp.VersionID = event.VersionID.Int64()
	}

	return resp
}